requirements.txt
streamlit
plotly
psycopg2-binary
pandas
time
psutil